/update-profile: fix ugly bug for Web users that deleted their profiles
introduced in 4df76d0db7b87cabbd714039546c05b3221169be

+++ pages.py
@@ -182,7 +182,7 @@ def update_profile(protocol, id):
                            authed_as=user.key.id())
         flash(f'Updating profile from {link}...')
 
-        if user.LABEL == 'web':
+        if user.status and user.LABEL == 'web':
             logger.info(f'Disabling web user {user.key.id()}')
             user.delete()
